requirements.txt
cx_Oracle
Pillow
Pandas
python-dotenv
schedule